chore: move developerUtils functions to developers/utils #627
