+++ src/controllers/developers/index.js
@@ -1,3 +1,12 @@
 import { getDeveloper, postDeveloper } from './developers';
+import { findDeveloper, FALLBACK_API_KEY, checkDeveloperAPIKey, fetchAPIKey, fetchRequestQuery } from './utils';
 
-export { getDeveloper, postDeveloper };
+export {
+  getDeveloper,
+  postDeveloper,
+  FALLBACK_API_KEY,
+  checkDeveloperAPIKey,
+  fetchAPIKey,
+  fetchRequestQuery,
+  findDeveloper,
+};
